Add test and fixtures
