+++ package.json
@@ -302,17 +302,20 @@
     "vscode:prepublish": "npm run compile",
     "compile": "tsc -p ./",
     "watch": "tsc -watch -p ./",
-    "test": "npm run compile && node ./node_modules/vscode/bin/test"
+    "test": "mocha -r ts-node/register test/unit/**/*.test.ts"
   },
   "devDependencies": {
+    "@types/chai": "^4.2.11",
+    "@types/jszip": "^3.1.7",
     "@types/mocha": "^2.2.48",
     "@types/node": "^8.10.59",
-    "@types/jszip": "^3.1.7",
-    "@types/vscode": "^1.42.0",
     "@types/parsimmon": "^1.10.1",
+    "@types/vscode": "^1.42.0",
+    "chai": "^4.2.0",
     "mocha": "^3.5.0",
-    "typescript": "^3.5.1",
-    "tslint": "^5.16.0"
+    "ts-node": "^8.8.1",
+    "tslint": "^5.16.0",
+    "typescript": "^3.5.1"
   },
   "dependencies": {
     "fast-xml-parser": "^3.16.0",
