+++ src/class_data/class_file_context.ts
@@ -1,4 +1,4 @@
-interface ClassFileContext {
+export interface ClassFileContext {
     path: string
     data: Buffer
     offset: number
