+++ test/unit/project_repository.test.ts
@@ -1,0 +1,7 @@
+import { ProjectRepository } from '../../src/project_repository'
+import { expect } from 'chai'
+import 'mocha'
+
+describe('Project Repository', () => {
+    let repo = new ProjectRepository()
+})
